LPS-137441 - Extract modal to a new file. Remove edit modal, will add it back later
